+++ src/transform/simfilefix.rs
@@ -9,52 +9,87 @@ use crate::transform::prelude::*;
 pub struct SimfileFix {
     pub from: BucketId,
     pub into: BucketId,
-    /// Which difficulties to output.
     /// Only one of each difficulty can be output for each gamemode, effectively limiting the
     /// amount of charts per song per gamemode to 6, tops.
     /// Damn good design.
-    ///
-    /// Defaults to the entire range of difficulties (`Beginner` - `Challenge`, `Edit`).
-    pub out_difficulties: Vec<Difficulty>,
-    /// Holds the difficulty number equivalent to each entry in `out_diffs`.
-    /// Used to map meters -> difficulty.
-    pub equivalent_meters: Vec<f64>,
+    pub fix_diffs: Option<FixDiffs>,
     /// Fix the stupid simfile format that doesn't support holds ending and another note starting
     /// at the same time.
     /// Pushes hold tails that are on the same beat and key as another note 1 microbeat backward.
     pub fix_holds: bool,
+    /// Whether to automatically merge input simfiles by music/gamemode, or to process each input
+    /// list individually.
+    ///
+    /// Defaults to `true`.
+    pub merge: bool,
 }
 impl Default for SimfileFix {
     fn default() -> Self {
-        use crate::simfile::Difficulty::*;
         Self {
             from: default(),
             into: default(),
-            out_difficulties: vec![Beginner, Easy, Medium, Hard, Challenge, Edit],
-            equivalent_meters: vec![1., 2., 3.5, 5., 6.5, 8.],
+            fix_diffs: Some(default()),
             fix_holds: true,
+            merge: true,
+        }
+    }
+}
+
+#[derive(Clone, Debug, Serialize, Deserialize)]
+#[serde(default)]
+pub struct FixDiffs {
+    /// The maximum amount of difficulties to output.
+    /// Having a value larger than the length of `diffs` makes no effect.
+    pub max: usize,
+    /// Which difficulties to output.
+    ///
+    /// Defaults to the entire range of difficulties (`Beginner` - `Challenge`, `Edit`).
+    pub diffs: Vec<Difficulty>,
+    /// Holds the difficulty number equivalent to each entry in `out_diffs`.
+    /// Used to map meters -> difficulty.
+    pub meters: Vec<f64>,
+}
+impl Default for FixDiffs {
+    fn default() -> Self {
+        use crate::simfile::Difficulty::*;
+        Self {
+            max: 6,
+            diffs: vec![Beginner, Easy, Medium, Hard, Challenge, Edit],
+            meters: vec![1., 2., 3.5, 5., 6.5, 8.],
         }
     }
 }
 
 impl Transform for SimfileFix {
     fn apply(&self, store: &mut SimfileStore) -> Result<()> {
-        let mut by_music_gamemode: HashMap<(PathBuf, Gamemode), Vec<Box<Simfile>>> = default();
-        store.get_each(&self.from, |_, sm| {
-            let list = by_music_gamemode
-                .entry((sm.music.clone().unwrap_or_default(), sm.gamemode))
-                .or_default();
-            list.push(sm);
-            Ok(())
-        })?;
-        for (_, mut list) in by_music_gamemode {
-            spread_difficulties(self, &mut list)?;
-            for sm in list.iter_mut() {
-                sm.fix_tails()?;
+        let process_list = |store: &mut SimfileStore, mut list: Vec<Box<Simfile>>| -> Result<()> {
+            if let Some(conf) = &self.fix_diffs {
+                spread_difficulties(conf, &mut list)?;
+            }
+            if self.fix_holds {
+                for sm in list.iter_mut() {
+                    sm.fix_tails()?;
+                }
             }
             store.put(&self.into, list);
+            Ok(())
+        };
+        if self.merge {
+            let mut by_music_gamemode: HashMap<(PathBuf, Gamemode), Vec<Box<Simfile>>> = default();
+            store.get_each(&self.from, |_, sm| {
+                let list = by_music_gamemode
+                    .entry((sm.music.clone().unwrap_or_default(), sm.gamemode))
+                    .or_default();
+                list.push(sm);
+                Ok(())
+            })?;
+            for (_, list) in by_music_gamemode {
+                process_list(store, list)?;
+            }
+            Ok(())
+        } else {
+            store.get(&self.from, process_list)
         }
-        Ok(())
     }
     fn buckets_mut<'a>(&'a mut self) -> BucketIter<'a> {
         Box::new(
@@ -65,12 +100,12 @@ impl Transform for SimfileFix {
 }
 
 /// There seems to be a max of 6 difficulties, so use them wisely and sort them.
-pub fn spread_difficulties(conf: &SimfileFix, simfiles: &mut Vec<Box<Simfile>>) -> Result<()> {
+pub fn spread_difficulties(conf: &FixDiffs, simfiles: &mut Vec<Box<Simfile>>) -> Result<()> {
     ensure!(
-        conf.out_difficulties.len() == conf.equivalent_meters.len(),
-        "equivalent_meters must have the same length as out_difficulties"
+        conf.diffs.len() == conf.meters.len(),
+        "meters must have the same length as diffs"
     );
-    if conf.out_difficulties.is_empty() {
+    if conf.diffs.is_empty() || conf.max <= 0 {
         simfiles.clear();
         return Ok(());
     }
@@ -93,7 +128,7 @@ pub fn spread_difficulties(conf: &SimfileFix, simfiles: &mut Vec<Box<Simfile>>) 
     trace!("    sorted difficulties: {:?}", order);
 
     //Remove difficulties, mantaining as much spread as possible
-    while order.len() > conf.out_difficulties.len() {
+    while order.len() > conf.diffs.len() || order.len() > conf.max {
         //Find the smallest gap
         let (mut smallest, _) = order
             .windows(2)
@@ -137,7 +172,7 @@ pub fn spread_difficulties(conf: &SimfileFix, simfiles: &mut Vec<Box<Simfile>>) 
         .iter()
         .map(|chart| {
             let (diff_idx, _diffnum) = conf
-                .equivalent_meters
+                .meters
                 .iter()
                 .enumerate()
                 .min_by_key(|(_i, &diffnum)| SortableFloat((diffnum - chart.difficulty_num).abs()))
@@ -159,7 +194,7 @@ pub fn spread_difficulties(conf: &SimfileFix, simfiles: &mut Vec<Box<Simfile>>) 
                     let mut occupied_if = difficulties[idx as usize];
                     let mut cost = 0.;
                     while occupied_if >= 0
-                        && occupied_if < conf.out_difficulties.len() as isize
+                        && occupied_if < conf.diffs.len() as isize
                         && idx >= 0
                         && idx < difficulties.len() as isize
                     {
@@ -171,7 +206,7 @@ pub fn spread_difficulties(conf: &SimfileFix, simfiles: &mut Vec<Box<Simfile>>) 
                             break;
                         }
                     }
-                    if occupied_if < 0 || occupied_if >= conf.out_difficulties.len() as isize {
+                    if occupied_if < 0 || occupied_if >= conf.diffs.len() as isize {
                         //Saturated. Max cost
                         9999.
                     } else {
@@ -199,7 +234,7 @@ pub fn spread_difficulties(conf: &SimfileFix, simfiles: &mut Vec<Box<Simfile>>) 
                     && idx < difficulties.len() as isize
                     && (difficulties[idx as usize] - set_to) * dir <= 0
                 {
-                    set_to = set_to.min(conf.out_difficulties.len() as isize - 1).max(0);
+                    set_to = set_to.min(conf.diffs.len() as isize - 1).max(0);
                     trace!(
                         "      moving difficulties[{}] == {} -> {}",
                         idx,
@@ -221,7 +256,7 @@ pub fn spread_difficulties(conf: &SimfileFix, simfiles: &mut Vec<Box<Simfile>>) 
 
     //Convert back from difficulty indices to actual difficulties
     for (chart, diff_idx) in simfiles.iter_mut().zip(difficulties) {
-        chart.difficulty = conf.out_difficulties[diff_idx as usize];
+        chart.difficulty = conf.diffs[diff_idx as usize];
         chart.difficulty_num = chart.difficulty_num.round();
     }
     trace!(